refactor: 로그인, 회원가입 공통된 부분이 많아 component/Auth로 합쳐서 관리
component/Login에 위치하던 footer는 common폴더로 이동
