+++ src/components/Auth/SignUpForm/validator.ts
