+++ src/components/Common/Footer/Footer.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
-import FooterRow from "components/Login/Footer/FooterRow";
-import InstagramLinks from "components/Login/Footer/InstagramLinks";
+import FooterRow from "components/Common/Footer/FooterRow";
+import InstagramLinks from "components/Common/Footer/InstagramLinks";
 
 const FooterContainer = styled.footer`
     display: flex;
