+++ src/components/Common/Footer/FooterRow.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import FooterTextPiece from "components/Login/Footer/FooterTextPiece";
-import FooterContentProps from "components/Login/Footer/FooterText.type";
+import FooterTextPiece from "components/Common/Footer/FooterTextPiece";
+import FooterContentProps from "components/Common/Footer/FooterText.type";
 
 const FlexRow = styled.div`
     display: flex;
