+++ src/components/Common/Footer/FooterText.type.ts
