+++ src/components/Common/Footer/FooterTextPiece.tsx
